+++ crates/zng-view/src/config/dconf.rs
@@ -2,7 +2,10 @@ use std::{io::BufRead as _, time::Duration};
 
 use zng_unit::{Rgba, TimeUnits as _};
 use zng_view_api::{
-    config::{AnimationsConfig, ColorScheme, ColorsConfig, FontAntiAliasing, KeyRepeatConfig, LocaleConfig, MultiClickConfig, TouchConfig},
+    config::{
+        AnimationsConfig, ChromeConfig, ColorScheme, ColorsConfig, FontAntiAliasing, KeyRepeatConfig, LocaleConfig, MultiClickConfig,
+        TouchConfig,
+    },
     Event,
 };
 
@@ -90,6 +93,15 @@ pub fn colors_config() -> ColorsConfig {
 pub fn locale_config() -> LocaleConfig {
     // sys_locale
     super::other::locale_config()
+}
+
+pub fn chrome_config() -> ChromeConfig {
+    let is_wayland = std::env::var("WAYLAND_DISPLAY").is_ok();
+    let is_gnome = std::env::var("XDG_CURRENT_DESKTOP").map_or(false, |val| val.contains("GNOME"));
+    ChromeConfig {
+        pref_custom: is_gnome,
+        provided: is_wayland,
+    }
 }
 
 fn on_change(key: &str, s: &crate::AppEventSender) {